UI enhancement, added delete and edit feature through context menu and added exit in File menu

+++ src/com/smy/todoList/Controller.java
@@ -2,8 +2,11 @@ package com.smy.todoList;
 
 import com.smy.todoList.datamodel.TodoData;
 import com.smy.todoList.datamodel.TodoItem;
+import javafx.application.Platform;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
+import javafx.collections.transformation.FilteredList;
+import javafx.collections.transformation.SortedList;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
@@ -18,7 +21,9 @@ import javafx.util.Callback;
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+import java.util.Comparator;
 import java.util.Optional;
+import java.util.function.Predicate;
 
 
 public class Controller {
@@ -37,6 +42,15 @@ public class Controller {
 
     @FXML
     private ContextMenu listContextMenu;
+
+    @FXML
+    ToggleButton toggleButton;
+
+    private FilteredList<TodoItem> filteredList;
+
+    private Predicate<TodoItem> unfilteredItems;
+    private Predicate<TodoItem> todayItems;
+
 
 
     public void initialize() {
@@ -63,6 +77,7 @@ public class Controller {
 
             // DO SOMETHING TO EDIT
 
+                showEditItemDialog();
 
 
             }
@@ -84,7 +99,33 @@ public class Controller {
             }
         });
 
-        listView.setItems(TodoData.getInstance().getTodoItemList());
+        unfilteredItems = new Predicate<TodoItem>() {
+          @Override
+            public boolean test(TodoItem todoItem) {
+              return true;
+          }
+        };
+
+        todayItems = new Predicate<TodoItem>() {
+            @Override
+            public boolean test(TodoItem todoItem) {
+                return todoItem.getDueDate().equals(LocalDate.now());
+            }
+        };
+
+        filteredList = new FilteredList<>(TodoData.getInstance().getTodoItemList(),unfilteredItems);
+
+        SortedList<TodoItem> sortedList = new SortedList<>(filteredList, new Comparator<TodoItem>() {
+
+            @Override
+            public int compare(TodoItem o1, TodoItem o2) {
+                return o1.getDueDate().compareTo(o2.getDueDate());
+            }
+        });
+
+
+
+        listView.setItems(sortedList);
         listView.getSelectionModel().selectFirst();
         listView.getSelectionModel().setSelectionMode(SelectionMode.SINGLE);
         detailView.editableProperty().setValue(false);
@@ -173,6 +214,55 @@ public class Controller {
 
         }
 
+    public void showEditItemDialog() {
+
+        Dialog<ButtonType> dialog = new Dialog<>();
+        dialog.initOwner(mainBorderPane.getScene().getWindow());
+        FXMLLoader fxmlLoader = new FXMLLoader();
+        fxmlLoader.setLocation(getClass().getResource("editItemDialog.fxml"));
+
+        dialog.setTitle("Edit New Item");
+        dialog.setHeaderText("This Edits Item");
+
+        try {
+            dialog.getDialogPane().setContent(fxmlLoader.load());
+
+        }   catch (IOException e) {
+            System.out.println(e.getMessage());
+            e.printStackTrace();
+            return;
+        }
+
+        DialogController dialogController = fxmlLoader.getController();
+
+        TodoItem selectedItem = listView.getSelectionModel().getSelectedItem();
+
+         dialogController.preEditItem(selectedItem);
+
+
+        dialog.getDialogPane().getButtonTypes().add(ButtonType.APPLY);
+        dialog.getDialogPane().getButtonTypes().add(ButtonType.CANCEL);
+
+        Optional<ButtonType> result = dialog.showAndWait();
+
+        if (result.isPresent() && result.get() == ButtonType.APPLY) {
+
+            System.out.println("Apply Pressed!");
+
+
+            TodoItem newTodoItem = dialogController.processResult();
+
+            TodoData.getInstance().getTodoItemList().remove(selectedItem);
+
+            listView.getSelectionModel().select(newTodoItem);
+
+
+        }
+        else
+            System.out.println("Cancel Pressed");
+
+    }
+
         private void  deleteItem(TodoItem item) {
 
         Alert deleteAlert = new Alert(Alert.AlertType.CONFIRMATION);
@@ -192,21 +282,44 @@ public class Controller {
         @FXML
         private void handleDeleteKeyPress(KeyEvent keyEvent) {
 
-            if (keyEvent.getCode() == KeyCode.DELETE) {
-                TodoItem item = listView.getSelectionModel().getSelectedItem();
+            TodoItem item = listView.getSelectionModel().getSelectedItem();
 
-                      deleteItem(item);
-            }
+            if (item != null)
+
+                if (keyEvent.getCode() == KeyCode.DELETE)
+
+                          deleteItem(item);
 
         }
 
+        public void handleFilterButton() {
 
-//    @FXML
-//    public void handleClickListView() {
-//
-//        TodoItem item = listView.getSelectionModel().getSelectedItem();
-//
-//
-//    }
+        TodoItem todoItem = listView.getSelectionModel().getSelectedItem();
+
+            if (toggleButton.isSelected()) {
+                filteredList.setPredicate(todayItems);
+
+                if (filteredList.isEmpty()) {
+                    detailView.setText("");
+                    dueDate.setText("");
+                }
+
+                else if (filteredList.contains(todoItem)) listView.getSelectionModel().select(todoItem);
+
+                else listView.getSelectionModel().selectFirst();
+            }
+            else {
+
+                filteredList.setPredicate(unfilteredItems);
+                listView.getSelectionModel().select(todoItem);
+            }
+
+
+        }
+
+    @FXML
+    public void exitProgram() {
+        Platform.exit();
+    }
 
 }
